+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.69](https://github.com/arez/arez-spytools/tree/v0.69) (2019-07-16)
 [Full Changelog](https://github.com/arez/arez-spytools/compare/v0.68...v0.69)