+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.com.google.elemental2` artifacts to version `2.26`.
+
 ### [v0.143](https://github.com/arez/arez/tree/v0.143) (2019-07-16)
 [Full Changelog](https://github.com/arez/arez/compare/v0.142...v0.143)